celery and redis pub-sub poc
