+++ EasyTask/manage.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
-"""Django's command-line utility for administrative Celery."""
+"""Django's command-line utility for administrative TaskProcessor."""
 import os
 import sys
 
 
 def main():
-    """Run administrative Celery."""
+    """Run administrative TaskProcessor."""
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'EasyTask.settings')
     try:
         from django.core.management import execute_from_command_line
